Move all writer-only tests from tests/unit_tests.rs to tests/writer.rs
Moved tests:
- test_new_xml_decl_full       -> declaration::full
- test_new_xml_decl_standalone -> declaration::standalone
- test_new_xml_decl_encoding   -> declaration::encoding
- test_new_xml_decl_version    -> declaration::version
- test_new_xml_decl_empty      -> declaration::empty
