+++ tests/unit_tests.rs
@@ -4,7 +4,7 @@ use std::str::from_utf8;
 
 use quick_xml::events::attributes::{AttrError, Attribute};
 use quick_xml::events::Event::*;
-use quick_xml::events::{BytesDecl, BytesEnd, BytesStart, BytesText};
+use quick_xml::events::{BytesEnd, BytesStart, BytesText};
 use quick_xml::name::QName;
 use quick_xml::reader::Reader;
 use quick_xml::writer::Writer;
@@ -286,84 +286,6 @@ fn test_write_attrs() -> Result<()> {
     assert_eq!(result, expected.as_bytes());
 
     Ok(())
-}
-
-#[test]
-fn test_new_xml_decl_full() {
-    let mut writer = Writer::new(Vec::new());
-    writer
-        .write_event(Decl(BytesDecl::new("1.2", Some("utf-X"), Some("yo"))))
-        .expect("writing xml decl should succeed");
-
-    let result = writer.into_inner();
-    assert_eq!(
-        String::from_utf8(result).expect("utf-8 output"),
-        "<?xml version=\"1.2\" encoding=\"utf-X\" standalone=\"yo\"?>",
-        "writer output (LHS)"
-    );
-}
-
-#[test]
-fn test_new_xml_decl_standalone() {
-    let mut writer = Writer::new(Vec::new());
-    writer
-        .write_event(Decl(BytesDecl::new("1.2", None, Some("yo"))))
-        .expect("writing xml decl should succeed");
-
-    let result = writer.into_inner();
-    assert_eq!(
-        String::from_utf8(result).expect("utf-8 output"),
-        "<?xml version=\"1.2\" standalone=\"yo\"?>",
-        "writer output (LHS)"
-    );
-}
-
-#[test]
-fn test_new_xml_decl_encoding() {
-    let mut writer = Writer::new(Vec::new());
-    writer
-        .write_event(Decl(BytesDecl::new("1.2", Some("utf-X"), None)))
-        .expect("writing xml decl should succeed");
-
-    let result = writer.into_inner();
-    assert_eq!(
-        String::from_utf8(result).expect("utf-8 output"),
-        "<?xml version=\"1.2\" encoding=\"utf-X\"?>",
-        "writer output (LHS)"
-    );
-}
-
-#[test]
-fn test_new_xml_decl_version() {
-    let mut writer = Writer::new(Vec::new());
-    writer
-        .write_event(Decl(BytesDecl::new("1.2", None, None)))
-        .expect("writing xml decl should succeed");
-
-    let result = writer.into_inner();
-    assert_eq!(
-        String::from_utf8(result).expect("utf-8 output"),
-        "<?xml version=\"1.2\"?>",
-        "writer output (LHS)"
-    );
-}
-
-/// This test ensures that empty XML declaration attribute values are not a problem.
-#[test]
-fn test_new_xml_decl_empty() {
-    let mut writer = Writer::new(Vec::new());
-    // An empty version should arguably be an error, but we don't expect anyone to actually supply
-    // an empty version.
-    writer
-        .write_event(Decl(BytesDecl::new("", Some(""), Some(""))))
-        .expect("writing xml decl should succeed");
-
-    let result = writer.into_inner();
-    assert_eq!(
-        String::from_utf8(result).expect("utf-8 output"),
-        "<?xml version=\"\" encoding=\"\" standalone=\"\"?>",
-        "writer output (LHS)"
-    );
 }
 
 #[test]
